backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jenz0000 <2025.jenz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
zappa = "^0.56.1"
python-dotenv = "^0.21.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
